Restructure as PyPI package, publish to TestPyPI
- Move all modules to src/telly_spelly/ layout
- Add pyproject.toml with hatchling build system
- Convert imports to relative for package compatibility
- Update README with pip/uvx installation options
- Published as telly-spelly 0.1.0 on TestPyPI
