--- __init__.py
@@ -1,1 +1,0 @@
-# Empty file to make the directory a Python package
