+++ src/telly_spelly/__init__.py
@@ -1,0 +1,3 @@
+"""Telly Spelly - Voice-to-text transcription for KDE Plasma using Whisper."""
+
+__version__ = "0.1.0"
