+++ src/telly_spelly/clipboard_manager.py
